% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_weights.R
\name{find_weights}
\alias{find_weights}
\title{Weighted log-rank test}
\usage{
find_weights(
  formula,
  data,
  wlr,
  t_star = NULL,
  s_star = NULL,
  rho = NULL,
  gamma = NULL,
  include_cens = TRUE
)
}
\arguments{
\item{formula}{Formula object. The response (on the left of the \code{~} operator) must be a survival object as returned by the \code{Surv} function.
The terms (on the right of the \code{~} operator) must include the treatment arm indicator, and additionally can include strata using the \code{strata} function.}

\item{data}{Data frame containing time-to-event data.}

\item{wlr}{Character string specifying type of weighted log-rank test. Either \code{"lr"} for a standard log-rank test, \code{"mw"} for a modestly-weighted log-rank test, or \code{"fh"} for the Fleming-Harrington rho-gamma family.}

\item{t_star}{Parameter \eqn{t*} in the modestly weighted (\code{"mw"}) test, see Details.}

\item{s_star}{Parameter \eqn{s*} in the modestly weighted (\code{"mw"}) test, see Details.}

\item{rho}{Parameter \eqn{\rho} in the Fleming-Harrington (\code{"fh"}) test, see Details.}

\item{gamma}{Parameter \eqn{\gamma} in the Fleming-Harrington (\code{"fh"}) test, see Details.}
}
\value{
List containing the outcome of the weighted log-rank test.
\code{u} is the test statistic U for the weighted log-rank test
\code{v_u} is the variance of test statistic U
\code{z} is the Z-score
\code{trt_group} indicates which of the treatment arms the test statistic U corresponds to
}
\description{
This function can perform two types of weighted log-rank test,
the modestly-weighted log-rank test and the Fleming-Harrington (\eqn{\rho},\eqn{\gamma}) test, in addition to the standard log-rank test.
}
\details{
Select which of the three tests to perform using argument \code{wlr}.
The output is calculated as outlined in \code{vignette("weighted_log_rank_tests", package="wlrt")}.
}
\examples{
library(wlrt)
set.seed(1)
sim_data <- sim_events_delay(
  n_c = 5,
  n_e = 5,
  delay_e = 6,
  lambda_c = log(2)/9,
  lambda_e_1 = log(2)/9,
  lambda_e_2 = log(2)/18,
  rec_period = 12,
  rec_power = 1,
  max_cal_t = 36
)
#example setting t_star
find_weights(formula=Surv(event_time,event_status)~group,
  data=sim_data,
  wlr="mw",
  t_star = 4
)
}
\references{
Magirr, D. (2021).
Non-proportional hazards in immuno-oncology: Is an old perspective needed?.
Pharmaceutical Statistics, 20(3), 512-527.
DOI: 10.1002/pst.2091

Magirr, D. and Burman, C.F., 2019.
Modestly weighted logrank tests.
Statistics in medicine, 38(20), 3782-3790.
}
