% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moderate_cross}
\alias{moderate_cross}
\title{Time-to-event RCT data set with moderate crossing of survival curves.}
\format{
A data frame with 328 rows and 3 variables:
\describe{
\item{time}{time to event / censoring}
\item{event}{observed event 1 / 0}
\item{arm}{treatment arm}
...
}
}
\usage{
moderate_cross
}
\description{
A synthetic data set based on an RCT with crossing survival curves.
}
\keyword{datasets}
