% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poplar}
\alias{poplar}
\title{Data from the POPLAR study
Biomarker information and outcomes for patients in the POPLAR study}
\format{
A data frame with 287 rows and 25 variables.
}
\source{
\url{https://www.nature.com/articles/s41591-018-0134-3}
}
\usage{
poplar
}
\description{
Data from the POPLAR study
Biomarker information and outcomes for patients in the POPLAR study
}
\keyword{datasets}
