% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_rec_times.R
\name{sim_rec_times}
\alias{sim_rec_times}
\title{Simulate recruitment times}
\usage{
sim_rec_times(
  rec_model,
  n = NULL,
  rec_power = NULL,
  rec_period = NULL,
  rec_lambda = NULL,
  rec_duration = NULL
)
}
\arguments{
\item{{rec_model}}{Character string specifying the type of recruitment model. Either the power model \code{"power"}
or piecewise exponential \code{"exp"}.}

\item{{n}}{Number of recruitment times to simulate}

\item{{rec_power}}{Parameter used to model recruitment according to power model, see Details.}

\item{{rec_period}}{Parameter used to model recruitment according to power model, see Details.}

\item{{rec_lambda}}{Parameter used to model recruitment according to piecewise exponential model, see Details.}

\item{{rec_duration}}{Parameter used to model recruitment according to piecewise exponential model, see Details.}
}
\value{

}
\description{
Simulate recruitment times from either a power model or a piecewise exponential model
}
\details{
Recruitment is modeled using either the power model or the piecewise exponential model.

The power model is defined as:
\eqn{P(recruited before T) = (T / rec_period) ^ rec_power}, where
\eqn{rec_period} is the time at the end of recruitment period, and \eqn{rec_power} controls the rate of recruitment.
}
\examples{
library(wlrt)
set.seed(1)
#power model
sim_rec_times(n=10,
  rec_model= "power",
  rec_period=20,
  rec_power=1)
#piecewise exponential model, where duration is set
sim_rec_times(rec_model="exp",
          rec_lambda=c(1,2),
          rec_duration=c(12,8))
#piecewise exponential model, where number of events is set
sim_rec_times(rec_model="exp",
          rec_lambda=c(1,2),
          rec_duration=c(12,8),
          n=10)
}
