% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_events_delay.R
\name{sim_events_delay}
\alias{sim_events_delay}
\title{Simulate survival data from a two-arm trial with a delayed separation of survival curves}
\usage{
sim_events_delay(
  n_c,
  n_e,
  delay_e,
  lambda_c,
  lambda_e_1,
  lambda_e_2,
  rec_period,
  rec_power,
  max_cal_t
)
}
\arguments{
\item{{n_c}}{Number of patients on control treatment.}

\item{{n_e}}{Number of patients on experimental treatment.}

\item{{delay_e}}{Length of first period of the experimental arm.}

\item{{lambda_c}}{Rate of exponential distribution  of the control arm.}

\item{{lambda_e_1}}{Rate of exponential distribution during first period of the experimental arm.}

\item{{lambda_e_2}}{Rate of exponential distribution during second period of the experimental arm.}

\item{{rec_period}}{Parameter used to model recruitment according to power model, see Details.}

\item{{rec_power}}{Parameter used to model recruitment according to power model, see Details.}

\item{{max_cal_t}}{Calendar time at which the trial ends, all observations are censored at this time.}
}
\value{
Data frame with columns \code{event_time}, \code{event_status} (\code{1} = event, \code{0} = censored), and treatment arm indicator \code{group}.
}
\description{
Simulate survival data from a two-arm trial with a delayed separation of survival curves.
Survival times on the control arm are simulated from an exponential distribution.
Survival times on the experimental arm are simulated from a two-piece exponential distribution.
}
\details{
Recruitment is modeled using the power model
\eqn{P(recruited before T) = (T / rec\_period) ^ rec\_power}, where
\eqn{rec_period} is the time at the end of recruitment period, and \eqn{rec_power} controls the rate of recruitment.

Survival times are simulated from an exponential distribution with rate parameter \eqn{\lambda},
\eqn{f(t)=\lambda exp(-\lambda t)}. This distribution has a median value of \eqn{log(2)/\lambda};
this can be a useful fact when setting the rates \code{lambda_c}, \code{lambda_e_1}, \code{lambda_e_2}.

All observations are censored at the calendar time defined in argument \code{max_cal_t}.
}
\examples{
library(wlrt)
set.seed(1)
sim_events_delay(
  n_c = 5,
  n_e = 5,
  delay_e = 6,
  lambda_c = log(2)/9,
  lambda_e_1 = log(2)/9,
  lambda_e_2 = log(2)/18,
  rec_period = 12,
  rec_power = 1,
  max_cal_t = 36
)
}
